Removed the return-type error and reactivated error
Removed useless cast to (size_t *) in stable_marriage

+++ contents/cooley_tukey/code/c/fft.c
@@ -6,7 +6,7 @@
 #include <time.h>
 #include <fftw3.h>
 
-void fft(double complex *x, int n) {
+void fft(double complex *x, size_t n) {
     double complex y[n];
     memset(y, 0, sizeof(y));
     fftw_plan p;
@@ -17,7 +17,7 @@ void fft(double complex *x, int n) {
     fftw_execute(p);
     fftw_destroy_plan(p);
 
-    for (int i = 0; i < n; ++i) {
+    for (size_t i = 0; i < n; ++i) {
         x[i] = y[i] / sqrt((double)n);
     }
 }
@@ -81,11 +81,11 @@ void iterative_cooley_tukey(double complex *X, size_t N) {
     bit_reverse(X, N);
 
     for (int i = 1; i <= log2((double)N); ++i) {
-        int stride = pow(2, i);
+        size_t stride = pow(2, i);
         double complex w = cexp(-2.0 * I * M_PI / stride);
         for (size_t j = 0; j < N; j += stride) {
             double complex v = 1.0;
-            for (int k = 0; k < stride / 2; ++k) {
+            for (size_t k = 0; k < stride / 2; ++k) {
                 X[k + j + stride / 2] = X[k + j] - v * X[k + j + stride / 2];
                 X[k + j] -= (X[k + j + stride / 2] - X[k + j]);
                 v *= w;
